Refactor requet setup codes
